move section class into it's own folder
